# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/organizations/{id}:
    get:
      operationId: OrganizationsController_getOrganization
      summary: Get organization details
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Organization details
        '404':
          description: Organization not found
      tags:
        - Organizations
      security:
        - bearer: []
    put:
      operationId: OrganizationsController_updateOrganization
      summary: Update organization settings
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationDto'
      responses:
        '200':
          description: Organization updated successfully
        '404':
          description: Organization not found
      tags:
        - Organizations
      security:
        - bearer: []
  /api/v1/organizations/{id}/stats:
    get:
      operationId: OrganizationsController_getOrganizationStats
      summary: Get organization statistics and usage
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Organization statistics
        '404':
          description: Organization not found
      tags:
        - Organizations
      security:
        - bearer: []
...
  /api/v1/membership-plans:
    post:
      operationId: MembershipPlansController_createMembershipPlan
      summary: Create a new membership plan
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMembershipPlanDto'
      responses:
        '201':
          description: Membership plan created successfully
        '400':
          description: Bad request
      tags:
        - Membership Plans
      security:
        - gym-id: []
        - bearer: []
    get:
      operationId: MembershipPlansController_getGymMembershipPlans
      summary: Get all membership plans for gym
      parameters:
        - name: activeOnly
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of membership plans
      tags:
        - Membership Plans
      security:
        - gym-id: []
        - bearer: []
  /api/v1/membership-plans/{id}:
    get:
      operationId: MembershipPlansController_getMembershipPlan
      summary: Get membership plan details
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Membership plan details
        '404':
          description: Membership plan not found
      tags:
        - Membership Plans
      security:
        - gym-id: []
        - bearer: []
    put:
      operationId: MembershipPlansController_updateMembershipPlan
      summary: Update membership plan
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMembershipPlanDto'
      responses:
        '200':
          description: Membership plan updated successfully
        '404':
          description: Membership plan not found
      tags:
        - Membership Plans
      security:
        - gym-id: []
        - bearer: []
    delete:
      operationId: MembershipPlansController_deleteMembershipPlan
      summary: Delete membership plan (soft delete)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Membership plan deleted successfully
        '400':
          description: Cannot delete plan with active contracts
        '404':
          description: Membership plan not found
      tags:
        - Membership Plans
      security:
        - gym-id: []
        - bearer: []
  /api/v1/membership-plans/{id}/stats:
    get:
      operationId: MembershipPlansController_getMembershipPlanStats
      summary: Get membership plan statistics
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Membership plan statistics
        '404':
          description: Membership plan not found
      tags:
        - Membership Plans
      security:
        - gym-id: []
        - bearer: []
...
  /api/v1/leads:
    post:
      operationId: LeadsController_createLead
      summary: Create a new lead from public catalog
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeadDto'
      responses:
        '201':
          description: Lead created successfully
        '400':
          description: Bad request
      tags:
        - Leads
    get:
      operationId: LeadsController_searchLeads
      summary: Search leads
      parameters:
        - name: status
          required: false
          in: query
          schema:
            enum:
              - NEW
              - CONTACTED
              - INTERESTED
              - CONVERTED
              - LOST
            type: string
        - name: search
          required: false
          in: query
          schema:
            example: john
            type: string
        - name: assignedToUserId
          required: false
          in: query
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
        - name: startDate
          required: false
          in: query
          schema:
            example: '2024-01-01'
            type: string
        - name: endDate
          required: false
          in: query
          schema:
            example: '2024-01-31'
            type: string
        - name: limit
          required: false
          in: query
          schema:
            example: '10'
            type: string
        - name: offset
          required: false
          in: query
          schema:
            example: '0'
            type: string
      responses:
        '200':
          description: List of leads
      tags:
        - Leads
      security:
        - gym-id: []
        - bearer: []
  /api/v1/leads/{id}:
    get:
      operationId: LeadsController_getLead
      summary: Get lead details
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Lead details
        '404':
          description: Lead not found
      tags:
        - Leads
      security:
        - bearer: []
    put:
      operationId: LeadsController_updateLead
      summary: Update lead status and information
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLeadDto'
      responses:
        '200':
          description: Lead updated successfully
        '404':
          description: Lead not found
      tags:
        - Leads
      security:
        - bearer: []
  /api/v1/leads/stats/gym:
    get:
      operationId: LeadsController_getLeadStats
      summary: Get lead statistics for gym
      parameters: []
      responses:
        '200':
          description: Lead statistics
      tags:
        - Leads
      security:
        - gym-id: []
        - bearer: []
  /api/v1/leads/{id}/convert:
    post:
      operationId: LeadsController_convertLead
      summary: Convert lead to client
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Lead converted successfully
        '400':
          description: Lead already converted
        '404':
          description: Lead not found
      tags:
        - Leads
      security:
        - bearer: []
...
info:
  title: GymSpace API
  description: The GymSpace Gym Management System API
  version: '1.0'
  contact:
    name: GymSpace API Support
    email: support@gymspace.com
  license:
    name: MIT
...
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.gymspace.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
    gym-id:
      type: apiKey
      in: header
      name: X-Gym-Id
  schemas:
...
    UpdateOrganizationDto:
      type: object
      properties:
        country:
          type: string
          example: US
        currency:
          type: string
          example: USD
        timezone:
          type: string
          example: America/New_York
        settings:
          type: object
          example:
            theme: light
            language: en
...
    CreateMembershipPlanDto:
      type: object
      properties:
        name:
          type: string
          example: Plan Mensual
        description:
          type: string
          example: Acceso completo al gimnasio durante un mes
        price:
          type: number
          example: 49.99
        currency:
          type: string
          example: USD
        durationMonths:
          type: number
          example: 1
          description: Duration in months
        termsAndConditions:
          type: string
          example: Terms and conditions text
        allowsCustomPricing:
          type: boolean
          example: false
        maxEvaluations:
          type: number
          example: 4
        includesAdvisor:
          type: boolean
          example: false
        showInCatalog:
          type: boolean
          example: true
        features:
          example:
            - Acceso completo
            - Clases grupales
            - Evaluación mensual
          type: array
          items:
            type: string
        status:
          type: string
          example: active
          description: Plan status
          enum:
            - active
            - inactive
            - archived
      required:
        - name
        - price
        - durationMonths
    UpdateMembershipPlanDto:
      type: object
      properties:
        name:
          type: string
          example: Plan Mensual
        description:
          type: string
          example: Acceso completo al gimnasio durante un mes
        price:
          type: number
          example: 49.99
        currency:
          type: string
          example: USD
        durationMonths:
          type: number
          example: 1
          description: Duration in months
        termsAndConditions:
          type: string
          example: Terms and conditions text
        allowsCustomPricing:
          type: boolean
          example: false
        maxEvaluations:
          type: number
          example: 4
        includesAdvisor:
          type: boolean
          example: false
        showInCatalog:
          type: boolean
          example: true
        features:
          example:
            - Acceso completo
            - Clases grupales
            - Evaluación mensual
          type: array
          items:
            type: string
        status:
          type: string
          example: active
          description: Plan status
          enum:
            - active
            - inactive
            - archived
        isActive:
          type: boolean
          example: true
...
    CreateLeadDto:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: john.doe@example.com
        phone:
          type: string
          example: '+1234567890'
        gymId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        message:
          type: string
          example: I would like to know more about membership plans
        source:
          type: string
          example: Website Contact Form
        metadata:
          type: object
          example:
            preferredContactTime: morning
            interests:
              - weight loss
              - muscle gain
      required:
        - name
        - email
        - phone
        - gymId
    UpdateLeadDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - NEW
            - CONTACTED
            - INTERESTED
            - CONVERTED
            - LOST
          example: CONTACTED
        notes:
          type: string
          example: Follow up scheduled for next week
        assignedToUserId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        metadata:
          type: object
          example:
            lastContactDate: '2024-01-15'
            nextFollowUp: '2024-01-22'
...